Move get_columns method into appropriate module for unix tty
